npbackup/requirements-win32.txt
command_runner>=1.7.0
cryptidy>=1.2.4
python-dateutil
ofunctions.logger_utils>=2.4.2
ofunctions.misc>=1.8.0
ofunctions.process>=2.1.0
ofunctions.threading>=2.2.0
ofunctions.platform>=1.5.3
ofunctions.random
ofunctions.requestor>=1.2.2
ofunctions.mailer>=1.3.0
#python-pidfile>=3.0.0
# pysimplegui 5 has gone commercial, let's switch to freesimplegui
# keep in mind that freesimplegui might higher required python version in the future
freesimplegui==5.2.0
requests
ruamel.yaml
psutil
pyyaml  # Required for python-i18n / i18nice which does not work with ruamel.yaml
# python-i18n
# Replaced python-i18n with a fork that prevents boolean keys from being interpreted
# Also fixes some portability issues (still boolean key name issues) encountered when compiling on Centos 7 and executing on Almalinux 9
# python-i18n@https://github.com/Krutyi-4el/python-i18n/archive/master.zip
# python-i18n @ git+https://github.com/Krutyi-4el/python-i18n.git@0.6.0#8999a0d380be8a08beed785e46fbb31dfc03c605
# Since PyPI and twine don't allow usage of direct references (git addresses)
# we'll use an inline version for now
i18nice>=0.6.2
packaging
pywin32; platform_system == "Windows"
imageio; platform_system == "Darwin"
ntplib>=0.4.0
# msgspec needs python 3.8+ and is not compatible with win 32-bit

